feat: Prepare project for debug
This initial task are to prepare the project to debug on Udacity Linux-Ubuntu Desktop. It adds the .vscode/launch.json and .vscode/tasks.json files with configurations to debug this project through cmake.  I've also added the .gitignore and .clang-format files and a new MakeFile on projct directory to ease project build, based on those present on previous Course Project: CppND-System-Monitor.